Feature/change api

+++ app/src/main/java/com/jangjh123/shallwegoforawalk/data/model/weather/HourlyWeather.kt
@@ -4,6 +4,6 @@ data class HourlyWeather(
     val temp: Int,
     val humidity: Int,
     val pop: Int,
-    val windSpeed: Float,
+    val windSpeed: Int,
     val icon: String
 )
